fix(websockets): websockets for ethernet adapter (#42)
* check for https to use wss connection
* fix(websocket): try only for 30s to connect
* the timer restarts on page reload
* fix(react): websocket via Ethernet

ScrabScrap previously used the IP address of the interface WLAN0
to generate a WebSocket address. This does not work if ScrabScrap
is connected via Ethernet.

Now the REACT application tries to establish a WebSocket connection (approx. 30s). 
If this works, this connection is preferred. As long as no connection can be 
established, file polling is used.

A configuration with a "websocket.js" file is no longer necessary.